moved trywithvariant to inference folder
