+++ ui/app/components/inference/TryWithVariantButton.tsx
